FirehoseRequestServer accepts a chaindb
